requirements.txt
# {# pkglts, pysetup
# requirements are managed by pkglts, do not edit this file at all
# edit .pkglts/pkg_cfg instead
# section pysetup

# install


# run examples


# test
pytest
pytest-mock


# compile doc
sphinx


# other dvlpt tools


# #}
path.py
six
nbformat
numpy
Cython
nbsphinx
networkx
graphviz
pydot
pyyaml
unyt
IPython
antlr4-python3-runtime == 4.8
